my-agent-app/azure_function_app/requirements.txt
azure-functions
azure-storage-file-share
openai
requests
pandas
tensorflow
numpy
